requirements.txt
asgiref==3.4.1
# autopep8==1.6.0
# backports.zoneinfo==0.2.1
coverage==6.3.2
Django==4.0
django-cors-headers==3.13.0
django-filter==21.1
djangorestframework==3.12.4
et-xmlfile==1.1.0
lml==0.1.0
openpyxl==3.0.9
Pillow==9.0.1
pycodestyle==2.8.0
pycryptodomex==3.12.0
pyexcel-io==0.6.6
pyexcel-xlsx==0.6.0
PyMySQL==1.0.2
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
